+++ examples/ESP32_Callbacks/ESP32_Callbacks.ino
@@ -3,7 +3,6 @@
 #define SerialMon Serial
 #define APPLEMIDI_DEBUG SerialMon
 #include <AppleMIDI.h>
-USING_NAMESPACE_APPLEMIDI
 
 #include "arduino_secrets.h"
 
@@ -12,9 +11,7 @@ bool isConnected = false;
 APPLEMIDI_CREATE_DEFAULTSESSION_INSTANCE();
 
 void setup() {
-  SerialMon.begin(115200);
-  while (!SerialMon);
-
+  DBG_SETUP(115200);
   DBG("Booting");
 
   WiFi.begin(SECRET_SSID, SECRET_PASS);
@@ -65,7 +62,7 @@ void loop() {
 // -----------------------------------------------------------------------------
 //
 // -----------------------------------------------------------------------------
-void OnAppleMidiConnected(const ssrc_t & ssrc, const char* name) {
+void OnAppleMidiConnected(const APPLEMIDI_NAMESPACE::ssrc_t & ssrc, const char* name) {
   isConnected = true;
   DBG(F("Connected to session"), name);
 }
@@ -73,7 +70,7 @@ void OnAppleMidiConnected(const ssrc_t & ssrc, const char* name) {
 // -----------------------------------------------------------------------------
 //
 // -----------------------------------------------------------------------------
-void OnAppleMidiDisconnected(const ssrc_t & ssrc) {
+void OnAppleMidiDisconnected(const APPLEMIDI_NAMESPACE::ssrc_t & ssrc) {
   isConnected = false;
 DBG(F("Disconnected"));
 }
@@ -81,7 +78,7 @@ DBG(F("Disconnected"));
 // -----------------------------------------------------------------------------
 //
 // -----------------------------------------------------------------------------
-void OnAppleMidiError(const ssrc_t& ssrc, int32_t err) {
+void OnAppleMidiError(const APPLEMIDI_NAMESPACE::ssrc_t& ssrc, int32_t err) {
   switch (err)
   {
     case Exception::BufferFullException:
