+++ examples/MKR_ETH_Bonjour/MKR_ETH_Bonjour.ino
@@ -4,7 +4,6 @@
 #define OPTIONAL_MDNS
 
 #include <AppleMIDI.h>
-USING_NAMESPACE_APPLEMIDI
 
 // Enter a MAC address for your controller below.
 // Newer Ethernet shields have a MAC address printed on a sticker on the shield
